Addressing PR comments
